+++ test/spec/modules/ucfunnelBidAdapter_spec.js
@@ -1,5 +1,5 @@
 import { expect } from 'chai';
-import Adapter from '../../../src/adapters/ucfunnel';
+import Adapter from 'src/adapters/ucfunnelBidAdapter';
 import adapterManager from 'src/adaptermanager';
 import bidManager from 'src/bidmanager';
 import CONSTANTS from 'src/constants.json';
